requirements/production.txt
django-environ==0.4.5
gunicorn==20.0.2
psycopg2-binary==2.8.4
django-storages==1.9.1
boto3==1.12.29

